requirements.txt
openai
playwright==1.41.2
google-generativeai==0.3.2
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2 
requests
beautifulsoup4
python-dotenv
fake-useragent
selenium
webdriver-manager 